requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
altair>=5.0.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
